Moved calculators to dedicated folder mortgage calculators and updated navigation
