+++ src/app/[locale]/_components/navigation.tsx
@@ -208,6 +208,24 @@ const more: More[] = [
     description: 'Careers',
     href: '/careers',
     icon: ChartPieIcon,
+  },
+    {
+    name: 'Mortgage Calculator',
+    description: 'Mortgage Calculator',
+    href: '/mortgage-calculator',
+    icon: ChartPieIcon,
+  },
+    {
+    name: 'Youngsters Program',
+    description: 'Youngsters Program',
+    href: '/psi-youngsters-program',
+    icon: ChartPieIcon,
+  },
+      {
+    name: 'Newsletter',
+    description: 'Newsletter',
+    href: '/newsletter',
+    icon: ChartPieIcon,
   }
 ];
