+++ src/app/[locale]/mortgage-calculator/AmortizationCalculator.tsx
