+++ src/app/[locale]/mortgage-calculator/HouseAffordabilityCalculator.tsx
