+++ src/app/[locale]/mortgage-calculator/MortgageCalculator.tsx
