+++ src/app/[locale]/mortgage-calculator/MortgageTabs.tsx
